+++ src/org/soulspace/clj/java/awt/events.clj
@@ -1,14 +1,17 @@
-;;
-;;   Copyright (c) Ludger Solbach. All rights reserved.
-;;   The use and distribution terms for this software are covered by the
-;;   Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0.php)
-;;   which can be found in the file license.txt at the root of this distribution.
-;;   By using this software in any fashion, you are agreeing to be bound by
-;;   the terms of this license.
-;;   You must not remove this notice, or any other, from this software.
-;;
+;;;;
+;;;;   Copyright (c) Ludger Solbach. All rights reserved.
+;;;;
+;;;;   The use and distribution terms for this software are covered by the
+;;;;   Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0.php)
+;;;;   which can be found in the file license.txt at the root of this distribution.
+;;;;   By using this software in any fashion, you are agreeing to be bound by
+;;;;   the terms of this license.
+;;;;
+;;;;   You must not remove this notice, or any other, from this software.
+;;;;
 
-(ns org.soulspace.clj.java.awt.event
+(ns org.soulspace.clj.java.awt.events
+  "Functions to create and add AWT event listeners."
   (:import [java.beans PropertyChangeListener]
            [java.awt.event
             ActionListener AdjustmentListener ComponentAdapter ComponentListener
@@ -19,9 +22,9 @@
             TextListener WindowAdapter WindowFocusListener WindowListener WindowStateListener]
            [java.awt.dnd DragSourceListener DragSourceAdapter DropTargetListener DropTargetAdapter]))
 
-;;
-;; Functions to create and add AWT event listeners
-;;
+;;;;
+;;;; Functions to create and add AWT event listeners
+;;;;
 
 (defn no-operation-fn
   ""
@@ -32,7 +35,8 @@
 
 (defmacro defadapter
   ""
-  [adapter args & fn-bindings])
+  [adapter args & fn-bindings]
+  )
 
 
 ; Listeners