+++ src/org/soulspace/clj/java/print.clj
@@ -1,14 +1,17 @@
-;;
-;;   Copyright (c) Ludger Solbach. All rights reserved.
-;;   The use and distribution terms for this software are covered by the
-;;   Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0.php)
-;;   which can be found in the file license.txt at the root of this distribution.
-;;   By using this software in any fashion, you are agreeing to be bound by
-;;   the terms of this license.
-;;   You must not remove this notice, or any other, from this software.
-;;
+;;;;
+;;;;   Copyright (c) Ludger Solbach. All rights reserved.
+;;;;
+;;;;   The use and distribution terms for this software are covered by the
+;;;;   Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0.php)
+;;;;   which can be found in the file license.txt at the root of this distribution.
+;;;;   By using this software in any fashion, you are agreeing to be bound by
+;;;;   the terms of this license.
+;;;;
+;;;;   You must not remove this notice, or any other, from this software.
+;;;;
 
 (ns org.soulspace.clj.java.print
+  "Functions for the Java Print API."
   (:import [java.awt Graphics]
            [java.awt.print PrinterJob PageFormat Printable]
            [java.io OutputStream]
@@ -18,10 +21,6 @@
                         StreamPrintService StreamPrintServiceFactory]
            [javax.print.attribute Attribute AttributeSet HashPrintRequestAttributeSet]
            [javax.print.attribute.standard MediaSize MediaSizeName]))
-
-;;
-;; Functions for the Java Print API
-;;
 
 (def doc-flavors {:byte-array        DocFlavor$BYTE_ARRAY
                   :char-array        DocFlavor$CHAR_ARRAY