+++ src/org/soulspace/clj/java/swing/core.clj
@@ -1,27 +1,30 @@
-;;
-;;   Copyright (c) Ludger Solbach. All rights reserved.
-;;   The use and distribution terms for this software are covered by the
-;;   Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0.php)
-;;   which can be found in the file license.txt at the root of this distribution.
-;;   By using this software in any fashion, you are agreeing to be bound by
-;;   the terms of this license.
-;;   You must not remove this notice, or any other, from this software.
-;;
+;;;;
+;;;;   Copyright (c) Ludger Solbach. All rights reserved.
+;;;;
+;;;;   The use and distribution terms for this software are covered by the
+;;;;   Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0.php)
+;;;;   which can be found in the file license.txt at the root of this distribution.
+;;;;   By using this software in any fashion, you are agreeing to be bound by
+;;;;   the terms of this license.
+;;;;
+;;;;   You must not remove this notice, or any other, from this software.
+;;;;
 
-(ns org.soulspace.clj.java.swing.swinglib
-  (:require [org.soulspace.clj.function :as fn]
-            [org.soulspace.clj.java.beans :as b]
-            [org.soulspace.clj.java.swing.constants :as sc])
-  (:import [java.awt CardLayout]
-           [javax.swing AbstractAction AbstractListModel Action BorderFactory ButtonGroup
+(ns org.soulspace.clj.java.swing.core
+  "Functions to create Swing user interfaces."
+  (:require [org.soulspace.clj.core :as fn]
+            [org.soulspace.clj.java.beans :as b])
+  (:import [java.awt CardLayout Event]
+           [java.awt.event ActionEvent]
+           [javax.swing Action AbstractAction AbstractListModel Action BorderFactory ButtonGroup
             ImageIcon InputVerifier
-            JButton JCheckBox JCheckBoxMenuItem JColorChooser JComboBox JDesktopPane JDialog
+            JButton JCheckBox JCheckBoxMenuItem JColorChooser JComboBox JComponent JDesktopPane JDialog
             JEditorPane JFileChooser JFormattedTextField JFrame JLabel JLayeredPane JList
             JMenu JMenuBar JMenuItem JOptionPane JPanel JPasswordField JPopupMenu JProgressBar
             JRadioButton JRadioButtonMenuItem JSeparator JScrollPane JSlider JSpinner
             JSplitPane JTabbedPane JTable JTextArea JTextField JTextPane JToggleButton JToolBar
             JTree JWindow
-            KeyStroke SwingConstants SwingUtilities UIManager]
+            KeyStroke ListSelectionModel SwingConstants SwingUtilities UIManager WindowConstants]
            [javax.swing.event AncestorListener CaretListener CellEditorListener ChangeListener DocumentListener
             HyperlinkListener InternalFrameListener ListDataListener ListSelectionListener
             MenuDragMouseListener MenuKeyListener MenuListener MouseInputListener PopupMenuListener
@@ -34,9 +37,104 @@
            [java.text DateFormat Format NumberFormat ParseException]
            [net.miginfocom.swing MigLayout]))
 
-;;
-;; Functions to create Swing UIs
-;;
+;;;
+;;; Constant maps
+;;;
+
+(def swing-keys
+  "Maps keywords to the constants in SwingConstants."
+  {:bottom     SwingConstants/BOTTOM
+   :center     SwingConstants/CENTER
+   :east       SwingConstants/EAST
+   :horizontal SwingConstants/HORIZONTAL
+   :leading    SwingConstants/LEADING
+   :left       SwingConstants/LEFT
+   :next       SwingConstants/NEXT
+   :north      SwingConstants/NORTH
+   :north-east SwingConstants/NORTH_EAST
+   :north-west SwingConstants/NORTH_WEST
+   :previous   SwingConstants/PREVIOUS
+   :right      SwingConstants/RIGHT
+   :south      SwingConstants/SOUTH
+   :south-east SwingConstants/SOUTH_EAST
+   :south-west SwingConstants/SOUTH_WEST
+   :top        SwingConstants/TOP
+   :trailing   SwingConstants/TRAILING
+   :vertical   SwingConstants/VERTICAL
+   :west       SwingConstants/WEST})
+
+(def window-keys
+  "Maps keywords to the constants in WindowConstants."
+  {:nothing WindowConstants/DO_NOTHING_ON_CLOSE
+   :dispose WindowConstants/DISPOSE_ON_CLOSE
+   :hide    WindowConstants/HIDE_ON_CLOSE
+   :exit    WindowConstants/EXIT_ON_CLOSE})
+
+(def focus-condition-keys
+  "Maps keywords to the constants in JComponent."
+  {:undefined                          JComponent/UNDEFINED_CONDITION ; used by some of the APIs to mean that no condition is defined.
+   :when-ancestor-of-focused-component JComponent/WHEN_ANCESTOR_OF_FOCUSED_COMPONENT ; used for registerKeyboardAction that means that the command should be invoked when the receiving component is an ancestor of the focused component or is itself the focused component.
+   :when-focused                       JComponent/WHEN_FOCUSED ; used for registerKeyboardAction that means that the command should be invoked when the component has the focus.
+   :when-in-focused-window             JComponent/WHEN_IN_FOCUSED_WINDOW}) ; used for registerKeyboardAction that means that the command should be invoked when the receiving component is in the window that has the focus or is itself the focused component.
+
+(def action-keys
+  "Maps keywords to the constants in Action."
+  {:name        Action/NAME
+   :accelerator Action/ACCELERATOR_KEY
+   :command-key Action/ACTION_COMMAND_KEY
+   :long-desc   Action/LONG_DESCRIPTION
+   :short-desc  Action/SHORT_DESCRIPTION
+   :mnemonic    Action/MNEMONIC_KEY
+   :icon        Action/SMALL_ICON})
+
+(def modifier-mask-keys
+  "Maps keywords to the constants in ActionEvent."
+  {:ctrl  ActionEvent/CTRL_MASK
+   :shift ActionEvent/CTRL_MASK
+   :alt   ActionEvent/ALT_MASK})
+
+(def list-selection-keys
+  "Maps keywords to the constants in ListSelectionModel."
+  {:single            ListSelectionModel/SINGLE_SELECTION
+   :single-interval   ListSelectionModel/SINGLE_INTERVAL_SELECTION
+   :multiple-interval ListSelectionModel/MULTIPLE_INTERVAL_SELECTION})
+
+(def commit-edit-keys
+  "Maps keywords to the constants in JFormattedTextField."
+  {:commit           JFormattedTextField/COMMIT
+   :commit-or-revert JFormattedTextField/COMMIT_OR_REVERT
+   :persist          JFormattedTextField/PERSIST
+   :revert           JFormattedTextField/REVERT})
+
+(def filechooser-keys
+  "Maps keywords to the constants in JFileChooser."
+  {:approve               JFileChooser/APPROVE_OPTION
+   :cancel                JFileChooser/CANCEL_OPTION
+   :error                 JFileChooser/ERROR_OPTION
+   :directories-only      JFileChooser/DIRECTORIES_ONLY
+   :files-only            JFileChooser/FILES_ONLY
+   :files-and-directories JFileChooser/FILES_AND_DIRECTORIES})
+
+; TODO use in message dialogs
+(def option-pane-message-keys
+  "Maps keywords to the constants in JOptionsPane."
+  {:error    JOptionPane/ERROR_MESSAGE
+   :warning  JOptionPane/WARNING_MESSAGE
+   :question JOptionPane/QUESTION_MESSAGE
+   :info     JOptionPane/INFORMATION_MESSAGE
+   :plain    JOptionPane/PLAIN_MESSAGE})
+
+(def option-pane-keys
+  "Maps keywords to the constants in JOptionsPane."
+  {:default       JOptionPane/DEFAULT_OPTION
+   :yes-no        JOptionPane/YES_NO_OPTION
+   :yes-no-cancel JOptionPane/YES_NO_CANCEL_OPTION
+   :ok-cancel     JOptionPane/OK_CANCEL_OPTION
+   :yes           JOptionPane/YES_OPTION
+   :no            JOptionPane/NO_OPTION
+   :cancel        JOptionPane/CANCEL_OPTION
+   :ok            JOptionPane/OK_OPTION
+   :closed        JOptionPane/CLOSED_OPTION})
 
 ; Helpers
 (defn init-swing
@@ -93,7 +191,7 @@
                   (actionPerformed [evt] (f evt)))]
      (doseq [[k v] args]
        ;(println (str k " : " (action-keys k) " -> " v))
-       (.putValue action (sc/action-keys k) v))
+       (.putValue action (action-keys k) v))
      (.setEnabled action true)
      action)))
 
@@ -102,7 +200,7 @@
   ([keycode]
    (KeyStroke/getKeyStroke keycode))
   ([keycode & modifiers]
-   (KeyStroke/getKeyStroke keycode (reduce + (map sc/modifier-mask-keys modifiers)))))
+   (KeyStroke/getKeyStroke keycode (reduce + (map modifier-mask-keys modifiers)))))
 
 (defn get-input-map
   "Returns the input map of the component."
@@ -463,12 +561,12 @@
 (defn horizontal-split-pane
   "Creates a horizontal split pane."
   [args items]
-  (init-swing (JSplitPane. (sc/swing-keys :horizontal)) args items))
+  (init-swing (JSplitPane. (swing-keys :horizontal)) args items))
 
 (defn vertical-split-pane
   "Creates a vertical split pane."
   [args items]
-  (init-swing (JSplitPane. (sc/swing-keys :vertical)) args items))
+  (init-swing (JSplitPane. (swing-keys :vertical)) args items))
 
 (defn scroll-pane
   "Creates a scroll pane."
@@ -581,10 +679,10 @@
    (JOptionPane/showMessageDialog nil text))
   ([text title type]
    (JOptionPane/showMessageDialog
-     nil text title (sc/option-pane-message-keys type)))
+     nil text title (option-pane-message-keys type)))
   ([text title type icon]
    (JOptionPane/showMessageDialog
-     nil text title (sc/option-pane-message-keys type) icon)))
+     nil text title (option-pane-message-keys type) icon)))
 
 (defn confirm-dialog
   "Creates a confirm dialog."
@@ -592,10 +690,10 @@
    (JOptionPane/showConfirmDialog nil text title options))
   ([text title options type]
    (JOptionPane/showConfirmDialog
-     nil text title options (sc/option-pane-message-keys type)))
+     nil text title options (option-pane-message-keys type)))
   ([text title options type icon]
    (JOptionPane/showConfirmDialog
-     nil text title options (sc/option-pane-message-keys type) icon)))
+     nil text title options (option-pane-message-keys type) icon)))
 
 (defn input-dialog
   "Creates an input dialog."
@@ -605,10 +703,10 @@
    (JOptionPane/showInputDialog nil text title))
   ([text title type]
    (JOptionPane/showInputDialog
-     nil text title (sc/option-pane-message-keys type)))
+     nil text title (option-pane-message-keys type)))
   ([text title type icon values initial]
    (JOptionPane/showInputDialog
-     nil text title (sc/option-pane-message-keys type) icon (into-array values) initial)))
+     nil text title (option-pane-message-keys type) icon (into-array values) initial)))
 
 (defn option-pane
   "Creates an option pane dialog."