+++ src/org/soulspace/clj/java/swing/events.clj
@@ -1,14 +1,17 @@
-;;
-;;   Copyright (c) Ludger Solbach. All rights reserved.
-;;   The use and distribution terms for this software are covered by the
-;;   Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0.php)
-;;   which can be found in the file license.txt at the root of this distribution.
-;;   By using this software in any fashion, you are agreeing to be bound by
-;;   the terms of this license.
-;;   You must not remove this notice, or any other, from this software.
-;;
+;;;;
+;;;;   Copyright (c) Ludger Solbach. All rights reserved.
+;;;;
+;;;;   The use and distribution terms for this software are covered by the
+;;;;   Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0.php)
+;;;;   which can be found in the file license.txt at the root of this distribution.
+;;;;   By using this software in any fashion, you are agreeing to be bound by
+;;;;   the terms of this license.
+;;;;
+;;;;   You must not remove this notice, or any other, from this software.
+;;;;
 
-(ns org.soulspace.clj.java.swing.event
+(ns org.soulspace.clj.java.swing.events
+  "Functions to create and add Swing event listeners."
   (:import [javax.swing.event AncestorListener CaretListener CellEditorListener ChangeListener DocumentListener
             HyperlinkListener InternalFrameListener ListDataListener ListSelectionListener
             MenuDragMouseListener MenuKeyListener MenuListener MouseInputAdapter MouseInputListener PopupMenuListener
@@ -16,14 +19,11 @@
             TreeExpansionListener TreeModelListener TreeSelectionListener TreeWillExpandListener
             UndoableEditListener]))
 
-;;
-;; Functions to create and add Swing event listeners
-;;
+;;;
+;;; Create listeners
+;;;
 
-;;
-;; Create listeners
 ;; TODO add adapters for multi listeners and add listener functions
-;;
 
 (defn ancestor-listener
   "Creates an ancestor listener. Calls function 'f' on ancestor updates."