+++ src/org/soulspace/clj/java/text.clj
@@ -1,20 +1,19 @@
-;;
-;;   Copyright (c) Ludger Solbach. All rights reserved.
-;;   The use and distribution terms for this software are covered by the
-;;   Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0.php)
-;;   which can be found in the file license.txt at the root of this distribution.
-;;   By using this software in any fashion, you are agreeing to be bound by
-;;   the terms of this license.
-;;   You must not remove this notice, or any other, from this software.
-;;
+;;;;
+;;;;   Copyright (c) Ludger Solbach. All rights reserved.
+;;;;
+;;;;   The use and distribution terms for this software are covered by the
+;;;;   Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0.php)
+;;;;   which can be found in the file license.txt at the root of this distribution.
+;;;;   By using this software in any fashion, you are agreeing to be bound by
+;;;;   the terms of this license.
+;;;;
+;;;;   You must not remove this notice, or any other, from this software.
+;;;;
 
 (ns org.soulspace.clj.java.text
+  "Functions to construct java.text.Format instances."
   (:require [org.soulspace.clj.java.beans :as b])
   (:import [java.text ChoiceFormat DateFormat DecimalFormat MessageFormat NumberFormat SimpleDateFormat]))
-
-;;
-;; Functions to construct java.text.Format instances
-;;
 
 (defn message-format
   "Creates a MessageFormat."