+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+Version 0.8.0
+-------------
+* moved namespace org.soulspace.clj.java.awt to org.soulspace.clj.java.awt.core
+* moved namespace org.soulspace.clj.java.swinglib to org.soulspace.clj.java.swing.core
+* integrated org.soulspace.clj.java.swing.constants into org.soulspace.clj.java.swing.core
+
 Version 0.7.3
 -------------
 * added a set-value method for formatted text fields